+++ src/main/java/com/solace/samples/jcsmp/patterns/DirectSubscriber.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package com.solace.samples.patterns;
+package com.solace.samples.jcsmp.patterns;
 
 import java.io.IOException;
