+++ src/main/java/com/solace/samples/jcsmp/samplers/RRDirectReplier.java
@@ -15,7 +15,7 @@
  * Copyright 2013-2020 Solace Corporation. All rights reserved.
  */
 
-package com.solace.samples.samplers;
+package com.solace.samples.jcsmp.samplers;
 
 import java.util.Map;
 import com.solace.samples.samplers.common.ArgParser;
