+++ src/main/java/com/solace/samples/jcsmp/samplers/RRGuaranteedRequester.java
@@ -17,7 +17,7 @@
  * Copyright 2013-2020 Solace Corporation. All rights reserved.
  */
 
-package com.solace.samples.samplers;
+package com.solace.samples.jcsmp.samplers;
 
 import java.util.Map;
 import com.solace.samples.samplers.common.ArgParser;
