+++ src/main/java/com/solace/samples/jcsmp/samplers/SecureSession.java
@@ -16,7 +16,7 @@
  * Copyright 2009-2020 Solace Corporation. All rights reserved.
  */
 
-package com.solace.samples.samplers;
+package com.solace.samples.jcsmp.samplers;
 
 import com.solace.samples.samplers.common.ArgParser;
 import com.solace.samples.samplers.common.SampleApp;
