+++ src/main/java/com/solace/samples/jcsmp/samplers/SimpleFlowToTopic.java
@@ -13,7 +13,7 @@
  * Copyright 2007-2020 Solace Corporation. All rights reserved.
  */
 
-package com.solace.samples.samplers;
+package com.solace.samples.jcsmp.samplers;
 
 import com.solace.samples.samplers.common.ArgParser;
 import com.solace.samples.samplers.common.SampleApp;
